+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- feat(databases): add fork endpoint
+
 ## [0.7.0] - 2026-07-14
 
 ### Changed